requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
distro==1.9.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.51.0
playsound==1.3.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.98
pywin32==306
redis==5.1.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
spotipy==2.24.0
srt==3.5.3
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
vosk==0.3.45
websockets==13.1
